fix: render usernames with nick pattern
When a nickname is updated for a user, and it isn't updated as a
GuildMember object, we fail to apply the nick pattern to the Matrix
nickname for the ghost user. By using the same logic as the
GetUserStateForGuildMember in GetUserStateForDiscordUser, the nickname
should now match the nickname pattern. Hopefully this will cut back on
nickname changes being echoed out in Matrix channels.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>